@@ -1,15 +1,16 @@
 {
   "name": "@emberwatch/ember-language-server",
   "description": "A language server implementation for Ember.js projects.",
-  "version": "0.2.1",
+  "version": "0.2.2",
   "contributors": [
+    "Aleksandr Kanunnikov <lifeart92@gmail.com>",
     "Tobias Bieniek <tobias.bieniek@gmail.com>",
     "Thomas Sauer <t-sauer@outlook.de>"
   ],
   "license": "MIT",
   "repository": {
     "type": "git",
-    "url": "https://github.com/emberwatch/ember-language-server.git"
+    "url": "https://github.com/lifeart/ember-language-server.git"
   },
   "engines": {
     "node": "*"
@@ -17,37 +18,72 @@
   "main": "lib/start-server.js",
   "typings": "lib/start-server.d.ts",
   "dependencies": {
-    "@glimmer/syntax": "^0.38.1",
-    "ast-types": "^0.11.3",
-    "babylon": "^6.18.0",
-    "fs-extra": "^5.0.0",
+    "@glimmer/syntax": "^0.43.0",
+    "ast-types": "^0.13.2",
+    "dag-map": "^2.0.2",
+    "ember-meta-explorer": "^0.1.0",
+    "extract-tagged-template-literals": "^1.0.1",
+    "fs-extra": "^8.1.0",
     "fuzzaldrin": "^2.1.0",
     "i": "^0.3.5",
-    "vscode-languageserver": "^4.1.2",
-    "vscode-uri": "^1.0.6",
-    "walk-sync": "^1.0.1"
+    "lodash": "^4.17.11",
+    "memoizee": "^0.4.14",
+    "rimraf": "^3.0.0",
+    "util": "^0.12.1",
+    "vscode-languageserver": "^5.2.1",
+    "vscode-uri": "^2.1.1",
+    "walk-sync": "^2.0.2"
   },
   "devDependencies": {
-    "@types/babylon": "^6.16.2",
     "@types/estree": "*",
     "@types/fuzzaldrin": "^2.1.1",
-    "@types/jest": "^23.3.13",
-    "@types/node": "^9.3.0",
-    "jest": "^23.6.0",
+    "@types/jest": "^24.0.21",
+    "@types/lodash": "^4.14.144",
+    "@types/memoizee": "^0.4.3",
+    "@types/node": "^12.12.5",
+    "@typescript-eslint/eslint-plugin": "^2.6.0",
+    "@typescript-eslint/parser": "^2.6.0",
+    "broccoli-test-helper": "^2.0.0",
+    "eslint": "^6.6.0",
+    "eslint-config-prettier": "^6.5.0",
+    "eslint-plugin-prettier": "^3.1.1",
+    "husky": "^3.0.9",
+    "jest": "^24.9.0",
     "lerna-changelog": "^0.8.2",
-    "ts-jest": "^22.4.3",
-    "tslint": "^5.9.1",
-    "typescript": "^2.8.1",
-    "vscode-jsonrpc": "^3.6.1",
-    "vscode-languageserver-protocol": "^3.7.1"
+    "lint-staged": "^9.4.2",
+    "prettier": "^1.18.2",
+    "ts-jest": "^24.1.0",
+    "typescript": "^3.6.4",
+    "vscode-jsonrpc": "^4.0.0",
+    "vscode-languageserver-protocol": "^3.14.1"
+  },
+  "husky": {
+    "hooks": {
+      "pre-commit": "lint-staged"
+    }
+  },
+  "lint-staged": {
+    "*.ts": [
+      "yarn lint --fix",
+      "git add"
+    ]
+  },
+  "prettier": {
+    "tabWidth": 2,
+    "semi": true,
+    "printWidth": 160,
+    "singleQuote": true,
+    "arrowParens": "always",
+    "parser": "typescript"
   },
   "scripts": {
     "changelog": "lerna-changelog",
-    "clean": "rm -rf lib/",
+    "clean": "rimraf lib/",
     "compile": "tsc -p .",
-    "lint": "tslint --project ./tsconfig.json",
+    "lint": "eslint ./src/**/*.ts",
     "prepublish": "yarn clean && yarn compile",
     "test": "jest",
+    "test:coverage": "jest --coverage",
     "watch": "tsc --watch -p ."
   }
 }
