@@ -1,180 +1,137 @@
-import { extname, join } from 'path';
-
-import {
-  CompletionItem,
-  CompletionItemKind,
-  TextDocumentPositionParams
-} from 'vscode-languageserver';
-
+import { CompletionItem, TextDocumentPositionParams } from 'vscode-languageserver';
 import Server from '../server';
 import ASTPath from '../glimmer-utils';
 import { toPosition } from '../estree-utils';
 import { filter } from 'fuzzaldrin';
-
-const { preprocess } = require('@glimmer/syntax');
-
-import {
-  emberBlockItems,
-  emberMustacheItems,
-  emberSubExpressionItems
-} from './ember-helpers';
-import uniqueBy from '../utils/unique-by';
+import { queryELSAddonsAPIChain } from './../utils/addon-api';
+import { preprocess } from '@glimmer/syntax';
 import { getExtension } from '../utils/file-extension';
+import { log } from '../utils/logger';
+import { searchAndExtractHbs } from 'extract-tagged-template-literals';
 
-const walkSync = require('walk-sync');
-
+const extensionsToProvideTemplateCompletions = ['.hbs', '.js', '.ts'];
+const PLACEHOLDER = 'ELSCompletionDummy';
 export default class TemplateCompletionProvider {
   constructor(private server: Server) {}
+  getTextForGuessing(originalText: string, offset: number, PLACEHOLDER: string) {
+    return originalText.slice(0, offset) + PLACEHOLDER + originalText.slice(offset);
+  }
+  async provideCompletions(params: TextDocumentPositionParams): Promise<CompletionItem[]> {
+    log('provideCompletions');
+    const ext = getExtension(params.textDocument);
 
-  provideCompletions(params: TextDocumentPositionParams): CompletionItem[] {
+    if (ext !== null && !extensionsToProvideTemplateCompletions.includes(ext)) {
+      return [];
+    }
+
     const uri = params.textDocument.uri;
-
-    if (getExtension(params.textDocument) !== '.hbs') {
-      return [];
-    }
-
     const project = this.server.projectRoots.projectForUri(uri);
-    if (!project) {
+    const document = this.server.documents.get(uri);
+    if (!project || !document) {
+      return [];
+    }
+    const { root } = project;
+    const offset = document.offsetAt(params.position);
+    const position = document.positionAt(offset);
+    const documentContent = document.getText();
+    const originalText = ext === '.hbs' ? documentContent : searchAndExtractHbs(documentContent);
+    log('originalText', originalText);
+    let normalPlaceholder: any = PLACEHOLDER;
+    let ast: any = {};
+
+    const cases = [
+      PLACEHOLDER + ' />',
+      PLACEHOLDER,
+      PLACEHOLDER + '"',
+      PLACEHOLDER + "'",
+      PLACEHOLDER + '}} />',
+      PLACEHOLDER + '"}}',
+      PLACEHOLDER + '}}',
+      PLACEHOLDER + '}}{{/' + PLACEHOLDER + '}}',
+      // {{#}} -> {{# + P}}{{/P + }}
+      PLACEHOLDER + '}}{{/' + PLACEHOLDER,
+      PLACEHOLDER + ')}}',
+      PLACEHOLDER + '))}}',
+      PLACEHOLDER + ')))}}'
+    ];
+
+    while (cases.length) {
+      normalPlaceholder = cases.shift();
+      try {
+        let validText = this.getTextForGuessing(originalText, offset, normalPlaceholder);
+        ast = preprocess(validText);
+        log('validText', validText);
+        break;
+      } catch (e) {
+        log('parsing-error', this.getTextForGuessing(originalText, offset, normalPlaceholder));
+        ast = null;
+      }
+    }
+    log('ast must exists');
+    if (ast === null) {
       return [];
     }
 
-    let document = this.server.documents.get(uri);
-    let offset = document.offsetAt(params.position);
-    let originalText = document.getText();
-    let text = originalText.slice(0, offset) + 'ELSCompletionDummy' + originalText.slice(offset);
-    let ast = preprocess(text);
-    let focusPath = ASTPath.toPosition(ast, toPosition(params.position));
+    const focusPath = ASTPath.toPosition(ast, toPosition(params.position));
+
     if (!focusPath) {
       return [];
     }
 
-    const { root } = project;
-    let completions: CompletionItem[] = [];
+    const completions: CompletionItem[] = await queryELSAddonsAPIChain(project.builtinProviders.completionProviders, root, {
+      focusPath,
+      textDocument: params.textDocument,
+      position: params.position,
+      results: [],
+      server: this.server,
+      type: 'template',
+      originalText
+    });
 
-    if (isMustachePath(focusPath)) {
-      completions.push(...listComponents(root));
-      completions.push(...listHelpers(root));
-      completions.push(...emberMustacheItems);
-    } else if (isBlockPath(focusPath)) {
-      completions.push(...listComponents(root));
-      completions.push(...emberBlockItems);
-    } else if (isSubExpressionPath(focusPath)) {
-      completions.push(...listHelpers(root));
-      completions.push(...emberSubExpressionItems);
-    } else if (isLinkToTarget(focusPath)) {
-      completions.push(...listRoutes(root));
+    const addonResults = await queryELSAddonsAPIChain(project.providers.completionProviders, root, {
+      focusPath,
+      textDocument: params.textDocument,
+      position: params.position,
+      results: completions,
+      server: this.server,
+      type: 'template'
+    });
+    const textPrefix = getTextPrefix(focusPath, normalPlaceholder);
+    const endCharacterPosition = position.character;
+    if (textPrefix.length) {
+      position.character -= textPrefix.length;
     }
-
-    return filter(completions, getTextPrefix(focusPath), { key: 'label' });
+    return filter(addonResults, textPrefix, {
+      key: 'label',
+      maxResults: 40
+    }).map((rawEl: CompletionItem) => {
+      const el = Object.assign({}, rawEl);
+      if (el.textEdit) {
+        return el;
+      }
+      let endPosition = {
+        line: position.line,
+        character: endCharacterPosition
+      };
+      const shouldFixContent = normalPlaceholder.includes('}}{{');
+      el.textEdit = {
+        newText: shouldFixContent
+          ? normalPlaceholder
+              .split(PLACEHOLDER)
+              .join(el.label)
+              .replace('}}{{', '}}\n  \n{{')
+          : el.label,
+        range: {
+          start: position,
+          end: endPosition
+        }
+      };
+      return el;
+    });
   }
 }
 
-function listComponents(root: string): CompletionItem[] {
-  const jsPaths = walkSync(join(root, 'app', 'components'), {
-    directories: false,
-    globs: ['**/*.js']
-  });
-  const hbsPaths = walkSync(join(root, 'app', 'templates', 'components'), {
-    directories: false,
-    globs: ['**/*.hbs']
-  });
-  const paths = [...jsPaths, ...hbsPaths];
-
-  const items = paths
-    .map((filePath: string) => {
-      return {
-        kind: CompletionItemKind.Class,
-        label: filePath.replace(extname(filePath), ''),
-        detail: 'component',
-      };
-    });
-
-  return uniqueBy(items, 'label');
-}
-
-function listHelpers(root: string): CompletionItem[] {
-  const paths = walkSync(join(root, 'app', 'helpers'), {
-    directories: false,
-    globs: ['**/*.js']
-  });
-
-  const items = paths
-    .map((filePath: string) => {
-      return {
-        kind: CompletionItemKind.Function,
-        label: filePath.replace(extname(filePath), ''),
-        detail: 'helper',
-      };
-    });
-
-  return uniqueBy(items, 'label');
-}
-
-function listRoutes(root: string): CompletionItem[] {
-  const paths = walkSync(join(root, 'app', 'routes'), {
-    directories: false,
-    globs: ['**/*.js']
-  });
-
-  const items = paths
-    .map((filePath: string) => {
-      const label = filePath
-        .replace(extname(filePath), '')
-        .replace(/\//g, '.');
-      return {
-        kind: CompletionItemKind.File,
-        label,
-        detail: 'route',
-      };
-    });
-
-  return uniqueBy(items, 'label');
-}
-
-function isMustachePath(path: ASTPath): boolean {
-  let node = path.node;
-  if (node.type !== 'PathExpression') { return false; }
-  let parent = path.parent;
-  if (!parent || parent.type !== 'MustacheStatement') { return false; }
-  return parent.path === node;
-}
-
-function isBlockPath(path: ASTPath): boolean {
-  let node = path.node;
-  if (node.type !== 'PathExpression') { return false; }
-  let parent = path.parent;
-  if (!parent || parent.type !== 'BlockStatement') { return false; }
-  return parent.path === node;
-}
-
-function isSubExpressionPath(path: ASTPath): boolean {
-  let node = path.node;
-  if (node.type !== 'PathExpression') { return false; }
-  let parent = path.parent;
-  if (!parent || parent.type !== 'SubExpression') { return false; }
-  return parent.path === node;
-}
-
-function isLinkToTarget(path: ASTPath): boolean {
-  return isInlineLinkToTarget(path) || isBlockLinkToTarget(path);
-}
-
-function isInlineLinkToTarget(path: ASTPath): boolean {
-  let node = path.node;
-  if (node.type !== 'StringLiteral') { return false; }
-  let parent = path.parent;
-  if (!parent || parent.type !== 'MustacheStatement') { return false; }
-  return parent.params[1] === node && parent.path.original === 'link-to';
-}
-
-function isBlockLinkToTarget(path: ASTPath): boolean {
-  let node = path.node;
-  if (node.type !== 'StringLiteral') { return false; }
-  let parent = path.parent;
-  if (!parent || parent.type !== 'BlockStatement') { return false; }
-  return parent.params[0] === node && parent.path.original === 'link-to';
-}
-
-function getTextPrefix({ node: { original = '' } }: ASTPath): string {
-  return original.replace('ELSCompletionDummy', '');
+function getTextPrefix({ node }: ASTPath, normalPlaceholder: string): string {
+  let target = node.original || node.tag || node.name || node.chars || '';
+  return target.replace(normalPlaceholder, '').replace(PLACEHOLDER, '');
 }
