@@ -1,10 +1,10 @@
-import { extname } from 'path';
+import * as path from 'path';
 import { TextDocumentIdentifier } from 'vscode-languageserver';
 import { uriToFilePath } from 'vscode-languageserver/lib/files';
 
 export function getExtension(textDocument: TextDocumentIdentifier): string | null {
   const filePath = uriToFilePath(textDocument.uri);
-  const ext = filePath ? extname(filePath) : '';
+  const ext = filePath ? path.extname(filePath) : '';
 
   if (ext === '.handlebars') {
     return '.hbs';
